requirements.txt
alembic==0.9.6
aniso8601==8.1.0
bcrypt==3.1.4
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==6.7
coverage==4.4.2
enum-compat==0.0.3
eventlet==0.21.0
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.3
Flask-Migrate==2.1.1
flask-restx==0.2.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
greenlet==1.0.0
gunicorn==19.7.1
idna==2.10
itsdangerous==0.24
Jinja2==2.11.2
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.1.1
marshmallow==3.10.0
psycopg2-binary==2.8.5
pycparser==2.18
PyJWT==1.5.3
python-dateutil==2.6.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2017.3
redis==3.5.3
requests==2.25.1
rq==1.7.0
selenium==3.8.1
six==1.12.0
SQLAlchemy==1.3.22
SQLAlchemy-Utils==0.36.8
urllib3==1.26.3
Werkzeug==0.15.5
